SBG Updates
Just a few changes to the code to check the header data instead of printing the whole dataset for our sanity checks. Otherwise, we need to test the code. Note: the code won't run unless you have file paths specified for the CSV files you want to open so whoever is going to run this for the presentation needs to make sure the filepath is correct or our code is broken from the beginning.